충남대 BE_최효은 Step1 - 엔티티 매핑 (#147)
* docs: step1 요구사항 정리

* feat: JPA 의존성 추가 및 설정 변경

* feat: 엔티티 클래스 변경

* feat: Repository 인터페이스 변경

* feat: 멤버 서비스 변경 및 메서드 추가

* feat: 서비스 계층 jpa 적용

* refactor: member 객체로 조회하는 방식으로 변경

* docs: 확인 체크

* test: DataJpaTest 적용 테스트 작성 및 관련 오류 해결

---------

+++ src/main/java/gift/controller/AdminProductController.java
@@ -31,7 +31,7 @@ public class AdminProductController {
         List<ProductResponseDto> products = productService.getProductList(page, size);
 
         // 총 페이지 계산용
-        int total = productService.getProductCount();
+        long total = productService.getProductCount();
         int totalPages = (int) Math.ceil((double) total / size);
 
         // 뷰로 전달
